Changed lib to use doubles not floats

+++ sphere.cpp
@@ -2,39 +2,39 @@
 
 namespace phys
 {
-    Sphere::Sphere(float radius,float mass,Vec3 position):
+    Sphere::Sphere(double radius,double mass,Vec3 position):
         m_radius{radius},
         m_mass{mass},
         m_position{position},
-        m_drag_coef{0.15f*3.14159f*radius*radius}
+        m_drag_coef{0.15*3.14159*radius*radius}
     {
     }
 
-    Sphere::Sphere(float radius,float mass,Vec3 position,Vec3 velocity):
+    Sphere::Sphere(double radius,double mass,Vec3 position,Vec3 velocity):
         m_radius{radius},
         m_mass{mass},
         m_position{position},
         m_velocity{velocity},
-        m_drag_coef{0.15f*3.14159f*radius*radius}
+        m_drag_coef{0.15*3.14159*radius*radius}
     {
     }
 
-    Sphere::Sphere(float radius,float mass,float c_r,Vec3 position):
+    Sphere::Sphere(double radius,double mass,double c_r,Vec3 position):
         m_radius{radius},
         m_mass{mass},
         m_position{position},
         m_coef_restitution{c_r},
-        m_drag_coef{0.15f*3.14159f*radius*radius}
+        m_drag_coef{0.15*3.14159*radius*radius}
     {
     }
 
-    Sphere::Sphere(float radius,float mass,float c_r,Vec3 position,Vec3 velocity):
+    Sphere::Sphere(double radius,double mass,double c_r,Vec3 position,Vec3 velocity):
         m_radius{radius},
         m_mass{mass},
         m_position{position},
         m_velocity{velocity},
         m_coef_restitution{c_r},
-        m_drag_coef{0.15f*3.14159f*radius*radius}
+        m_drag_coef{0.15*3.14159*radius*radius}
     {
     }
 
@@ -46,28 +46,28 @@ namespace phys
     {
         if (m_position.getX() > box->max_x-m_radius || m_position.getX() < box->min_x+m_radius)
         {
-            float dist_out;
+            double dist_out;
             dist_out = (m_position.getX()>0) ? dist_out=m_position.getX()+m_radius-box->max_x : dist_out=m_position.getX()-m_radius-box->min_x;
-            m_position -= Vec3{dist_out,0.f,0.f};
-            m_velocity *= Vec3{-1.f,1.f,1.f}*m_coef_restitution;
+            m_position -= Vec3{dist_out,0,0};
+            m_velocity *= Vec3{-1,1,1}*m_coef_restitution;
         }
         if (m_position.getY() > box->max_y-m_radius || m_position.getY() < box->min_y+m_radius)
         {
-            float dist_out;
+            double dist_out;
             dist_out = (m_position.getY()>0) ? dist_out=m_position.getY()+m_radius-box->max_y : dist_out=m_position.getY()-m_radius-box->min_y;
-            m_position -= Vec3{0.f,dist_out,0.f};
-            m_velocity *= Vec3{1.f,-1.f,1.f}*m_coef_restitution;
+            m_position -= Vec3{0,dist_out,0};
+            m_velocity *= Vec3{1,-1,1}*m_coef_restitution;
         }
         if (m_position.getZ() > box->max_z-m_radius || m_position.getZ() < box->min_z+m_radius)
         {
-            float dist_out;
+            double dist_out;
             dist_out = (m_position.getZ()>0) ? dist_out=m_position.getZ()+m_radius-box->max_z : dist_out=m_position.getZ()-m_radius-box->min_z;
-            m_position -= Vec3{0.f,0.f,dist_out};
-            m_velocity *= Vec3{1.f,1.f,-1.f}*m_coef_restitution;
+            m_position -= Vec3{0,0,dist_out};
+            m_velocity *= Vec3{1,1,-1}*m_coef_restitution;
         }
     }
 
-    void Sphere::update(float dt,Vec3 acceleration,Boundary* box)
+    void Sphere::update(double dt,Vec3 acceleration,Boundary* box)
     {
         m_velocity = m_velocity + acceleration*dt;
         m_position = m_position + m_velocity*dt;
@@ -86,12 +86,12 @@ namespace phys
         return m_velocity;
     }
 
-    float Sphere::getDragCoef() const
+    double Sphere::getDragCoef() const
     {
         return m_drag_coef;
     }
 
-    float Sphere::getRadius() const
+    double Sphere::getRadius() const
     {
         return m_radius;
     }
@@ -100,17 +100,17 @@ namespace phys
     {
         if (box == nullptr)
             return false;
-        if (m_position.getX() >= box->max_x-m_radius-0.1f || m_position.getX() <= box->min_x+m_radius+0.1f)
+        if (m_position.getX() >= box->max_x-m_radius-0.1 || m_position.getX() <= box->min_x+m_radius+0.1)
             return true;
-        else if (m_position.getY() >= box->max_y-m_radius-0.1f || m_position.getY() <= box->min_y+m_radius+0.1f)
+        else if (m_position.getY() >= box->max_y-m_radius-0.1 || m_position.getY() <= box->min_y+m_radius+0.1)
             return true;
-        else if (m_position.getZ() >= box->max_z-m_radius-0.1f || m_position.getZ() <= box->min_z+m_radius+0.1f)
+        else if (m_position.getZ() >= box->max_z-m_radius-0.1 || m_position.getZ() <= box->min_z+m_radius+0.1)
             return true;
         else
             return false;
     }
 
-    float Sphere::getMass() const
+    double Sphere::getMass() const
     {
         return m_mass;
     }
